backend/requirements.txt
Flask==2.3.2
flask_jwt_extended==4.4.4
pymongo==4.5.0
python-dotenv==1.0.0
Flask-CORS==5.0.0
werkzeug==2.3.4
